EBR_engine_simulation 2.0.ino
In this version I have done away with the DACs using the wave table and wrote some basic math equations to generate the proper waveforms. I have also incorporated the serial monitor input. This allows users to input a frequency of their choice, to change the RPM and speed at runtime. Anything below 1000 will be read as a speed frequency and anything above 1000 will be read as an RPM frequency.